% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{question}
\alias{question}
\title{Print questiontext}
\usage{
question(data, variable)
}
\arguments{
\item{data}{Specify the dataset that was generated with read_neps function}

\item{variable}{Specify the variable for which the question text should be printed. Argument can be a string or an unquoted variable name.}
}
\value{
A String.
}
\description{
`question()` is a convenient wrapper for the attr() function. It prints the attached questiontext from specified variable in the variable argument to the console.
}
\examples{
# Example with NEPS SC6 semantic structures spGap file
path <- system.file("extdata", "SC6_spGap_S_15-0-0.dta", package = "nepstools")
df_neps <- read_neps(path, english = TRUE, col_select = c("ID_t", "ts2912m"))

question(df_neps, "ts2912m")

}
