% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neps.R
\name{read_exp_fields}
\alias{read_exp_fields}
\title{A helper to extract meta info from data}
\usage{
read_exp_fields(datapath, cols = NULL, attr_type = NULL, only_value = F)
}
\description{
A helper to extract meta info from data
}
\keyword{internal}
