% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neps.R
\name{read_neps}
\alias{read_neps}
\title{Function, that reads neps data and attracts available meta infos to each variable. Select variables with col_select, specify language: German (default) or English and lastly specify if only most important meta info (default) should be attracted or all of it.}
\usage{
read_neps(
  datasetpath,
  col_select = NULL,
  english = F,
  compact_meta = T,
  charren = F
)
}
\arguments{
\item{datasetpath}{A datapath to a NEPS data file}

\item{col_select}{Specify variables that will be included in the dataset. If NULL, data is being loaded with all variables available}

\item{english}{If TRUE, dataset will  be loaded with english variable names, value labels and meta info}

\item{compact_meta}{If TRUE, only a selection of important meta informattion will be added to the data: Questiontexts, Interviewertexts, Harmonization Rules and instrument variable names}

\item{charren}{If TRUE, instrument variable names will replace normal variable names}
}
\value{
A Dataframe.
}
\description{
Function, that reads neps data and attracts available meta infos to each variable. Select variables with col_select, specify language: German (default) or English and lastly specify if only most important meta info (default) should be attracted or all of it.
}
\examples{
read_neps(path_to_data, col_select = c("var1","var2"), english = T)
}
