% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_season_codes}
\alias{replace_season_codes}
\title{Replace season codes with months}
\usage{
replace_season_codes(
  data,
  vars = NULL,
  values_to_replace = c(21, 24, 27, 30, 32)
)
}
\arguments{
\item{data}{A dataset to apply the function to.}

\item{vars}{Specify vars where season codes should be replaced with months. Optional, if set to NULL, all vars are being taken into account. The function than only select variables that have the label "month" or "monat". It is recommended to specify the variables to which the function should be applied.}

\item{values_to_replace}{These are the season codes. Usually doesnt need to be modified.}
}
\value{
A Dataframe.
}
\description{
`replace_season_codes()` replaces NEPS specific season codes in date variables to standard month codes. For example code 27: "Summer" will be replaced with Code 7: "July".
}
\examples{
# Example with NEPS SC6 semantic structures spGap file
path <- system.file("extdata", "SC6_spGap_S_15-0-0.dta", package = "nepstools")
df_neps <- read_neps(path, english = TRUE, col_select = c("ID_t", "ts2912m"))

# create some artificial datapoints
artificial_datapoints <- data.frame(
 ID_t = c(1, 2, 3, 24, 5),
 ts2912m = c(-97, 24, 5, 32, 30))

# add these artificial datapoints to the empty neps dataset
df_neps <- rbind(df_neps, artificial_datapoints)
print(df_neps)

df_neps_replaced <- replace_season_codes(df_neps)
print(df_neps_replaced)

}
