% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_values_with_na}
\alias{replace_values_with_na}
\title{Set missings in dataframe to NA}
\usage{
replace_values_with_na(
  data,
  vars = NULL,
  values_to_replace = c(seq(-99, -90), seq(-56, -51), seq(-29, -22))
)
}
\arguments{
\item{data}{A dataset to apply the function to.}

\item{vars}{Specify vars where missing values should be replaced with NA. If set to NULL, all variables will be used.}

\item{values_to_replace}{Specify values that should be replaced with NA. Default are all standard NEPS missing codes.}
}
\value{
A dataframe.
}
\description{
Set missings in dataframe to NA
}
