% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neps.R
\name{switch_var_names}
\alias{switch_var_names}
\title{A Helper to switch variable names in NEPS data to original instrument variable names. This feature is inspired by stata nepstools "charren" function. Background: During the data preparation process for the NEPS SUFs, the original variable names from the survey instruments are replaced with SUF-specific variable names. However, for some data users it may be useful to get the original instrument names. This function provides a quick way to do this, utilizing the metadata "NEPS_alias" stored within the datasets when being loaded with read_neps function.}
\usage{
switch_var_names(data)
}
\description{
A Helper to switch variable names in NEPS data to original instrument variable names. This feature is inspired by stata nepstools "charren" function. Background: During the data preparation process for the NEPS SUFs, the original variable names from the survey instruments are replaced with SUF-specific variable names. However, for some data users it may be useful to get the original instrument names. This function provides a quick way to do this, utilizing the metadata "NEPS_alias" stored within the datasets when being loaded with read_neps function.
}
\keyword{internal}
