% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{expand}
\alias{expand}
\title{Expand data}
\usage{
expand(data, duration)
}
\arguments{
\item{data}{A dataset that will be expanded.}

\item{duration}{Specify the variable that will be used for expanding the data, usually this is a duration (of episode) variable. It needs to be a vector with only valid non-negative numbers and no NAs.}
}
\value{
A Dataframe.
}
\description{
`expand()` duplicates rows by a variable specified in the duration argument. It is inspired by statas expand function.
}
\examples{
df <- data.frame(id = 1:3, duration = c(2, 1, 3))
expand(df, duration)

}
