% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{expand}
\alias{expand}
\title{Expand episode data to monthly structure}
\usage{
expand(data, duration)
}
\arguments{
\item{data}{A dataset that will be expanded.}

\item{duration}{Specify the variable that will be used for expanding the data, usually this is a duration (of episode) variable. It needs to be a vector with only valid non-negative numbers and no NAs.}
}
\value{
A Dataframe.
}
\description{
Expand episode data to monthly structure
}
