% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{question}
\alias{question}
\title{Print questiontext}
\usage{
question(data, variable)
}
\arguments{
\item{data}{Specify the dataset that was generated with read_neps function}

\item{variable}{Specify the variable for which the question text should be printed.}
}
\value{
A String.
}
\description{
`question()` is a convenient wrapper for the attr() function. It prints the attached questiontext from specified variable in the variable argument to the console.
}
\examples{
df <- data.frame(id = 1, dur = 1)
attr(df$id, "NEPS_questiontext_de") <- "example question?"
question(df, "id")

}
