% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{question}
\alias{question}
\title{Prints questiontext of variable to console}
\usage{
question(data, variable)
}
\arguments{
\item{data}{Specify the dataset that was generated with read_neps function}

\item{variable}{Specify the variable for which the question text should be printed.}
}
\value{
A String.
}
\description{
Prints questiontext of variable to console
}
\examples{
df <- data.frame(id = 1, dur = 1)
attr(df$id, "NEPS_questiontext_de") <- "example question?"
question(df, "id")

}
