% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neps.R
\name{read_neps}
\alias{read_neps}
\title{Read NEPS SUF files in .dta format}
\usage{
read_neps(
  datasetpath,
  col_select = NULL,
  row_select = Inf,
  english = FALSE,
  compact_meta = TRUE,
  charren = FALSE
)
}
\arguments{
\item{datasetpath}{A datapath to a NEPS data file}

\item{col_select}{Specify variables that will be included in the dataset. If set to NULL, data is being loaded with all available variables.}

\item{row_select}{Specify maximum number of rows to read.}

\item{english}{If set to TRUE, the dataset will be loaded with English variable and value labels, and metadata. If set to FALSE, German labels and metadata will be used instead.}

\item{compact_meta}{If set to TRUE, only a selection of important metadata will be added to the data, including question texts, interviewer texts, harmonization rules, and instrument variable names. If set to FALSE, all available metadata will be included.}

\item{charren}{If set to TRUE, instrument variable names will replace the standard variable names. If set to FALSE, the standard variable names will be retained.}
}
\value{
A Dataframe.
}
\description{
`read_neps()` reads NEPS SUF data in .dta file format and attracts available meta infos to each variable. Select variables with col_select, specify language: German (default) or English and lastly specify if only most important meta info (default) should be attracted or all of it.
}
\examples{
path <- system.file("extdata", "SC6_spGap_S_15-0-0.dta", package = "nepstools")
df <- read_neps(path, english = TRUE, charren = TRUE)
print(names(df))
attr(df$luendm, "NEPS_questiontext_")

}
