% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_season_codes}
\alias{replace_season_codes}
\title{Replace season codes with corresponding months}
\usage{
replace_season_codes(
  data,
  vars = NULL,
  values_to_replace = c(21, 24, 27, 30, 32)
)
}
\arguments{
\item{data}{A dataset to apply the function to.}

\item{vars}{Specify vars where season codes should be replaced with months. Optional, if set to NULL, all vars are being taken into account, which can lead to problems in case of non-date variables.}

\item{values_to_replace}{These are the season codes. Usually doesnt need to be modified.}
}
\value{
A Dataframe.
}
\description{
Replace season codes with corresponding months
}
