% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_season_codes}
\alias{replace_season_codes}
\title{Replace season codes with months}
\usage{
replace_season_codes(
  data,
  vars = NULL,
  values_to_replace = c(21, 24, 27, 30, 32)
)
}
\arguments{
\item{data}{A dataset to apply the function to.}

\item{vars}{Specify vars where season codes should be replaced with months. Optional, if set to NULL, all vars are being taken into account, which can lead to problems in case of non-date variables.}

\item{values_to_replace}{These are the season codes. Usually doesnt need to be modified.}
}
\value{
A Dataframe.
}
\description{
`replace_season_codes()` replaces NEPS specific season codes in date variables to standard month codes. For example code 27: "Summer" will be replaced with Code 7: "July".
}
\examples{
df <- data.frame(
  month_var = c(21, 22, 24, 25, 27),
  other_var = c(1, 2, 3, 4, 5)
)
attr(df$month_var, "label") <- "Month of survey"

replace_season_codes(df)

}
